refactor: (#740) 최신 알림 읽기 query hook 오타수정
