+++ frontend/src/hooks/query/user/useReadLatestAlarm.ts
